+++ src/index.js
@@ -1,1 +1,1 @@
-export { handler, trigger, hierarchyTrigger, nodesTrigger, labelsTrigger } from './be';
+export { handler, keywordTrigger, hierarchyTrigger, nodesTrigger, labelsTrigger } from './be';
